profile CRUD

+++ Fake-Reddit-Spring/src/main/java/com/ga/entity/Profile.java
@@ -30,7 +30,7 @@ public class Profile {
 //		this.address = address;
 //	}
 
-	public int getProfileId() {
+	public Long getProfileId() {
 		return profileId;
 	}
